+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.6 (????-??-??)
+
+* Added `ports` keyword argument to `launch` method for `ROSWire`, allowing
+  users to specify an optional container-host port mapping.
+
+
 # 0.0.5 (2019-08-28)
 
 * Created initial CHANGELOG.